requirements.txt
numpy==1.24.3
pytest==7.4.0
PyQt5==5.15.9
Pillow==10.0.0 